Refactoring: Html Sanitizer distinct from email fetcher
